+++ packages/sitecore-jss-nextjs/src/middleware/redirects-middleware.ts
@@ -82,29 +82,38 @@ export class RedirectsMiddleware extends MiddlewareBase {
     req: NextRequest,
     siteName: string
   ): Promise<RedirectResult | undefined> {
-    const { pathname: targetURL, search: targetQS = '' } = this.normalizeUrl(req.nextUrl.clone());
+    const { pathname: incomingURL, search: incomingQS = '' } = this.normalizeUrl(
+      req.nextUrl.clone()
+    );
     const locale = this.getLanguage(req);
-    const normalizedPath = targetURL.replace(/\/*$/gi, '');
+    const normalizedPath = incomingURL.replace(/\/*$/gi, '');
     const redirects = await this.redirectsService.fetchRedirects(siteName);
     const language = this.getLanguage(req);
     const modifyRedirects = structuredClone(redirects);
     let matchedQueryString: string | undefined;
+    const localePath = `/${locale}${normalizedPath}`.toLowerCase();
 
     return modifyRedirects.length
       ? modifyRedirects.find((redirect: RedirectResult) => {
+          // process static URL (non-regex) rules
           if (isRegexOrUrl(redirect.pattern) === 'url') {
             const [patternPath, patternQS] = redirect.pattern.endsWith('/')
-              ? redirect.pattern.slice(0, -1).split('?')
-              : redirect.pattern.split('?');
+              ? redirect.pattern
+                  .toLowerCase()
+                  .slice(0, -1)
+                  .split('?')
+              : redirect.pattern.toLowerCase().split('?');
             return (
-              (patternPath === normalizedPath || patternPath === `/${locale}${normalizedPath}`) &&
+              (patternPath === localePath || patternPath === normalizedPath) &&
               (!patternQS ||
                 areURLSearchParamsEqual(
                   new URLSearchParams(patternQS),
-                  new URLSearchParams(targetQS)
+                  new URLSearchParams(incomingQS)
                 ))
             );
           }
+
+          // process regex rules
 
           // Modify the redirect pattern to ignore the language prefix in the path
           // And escapes non-special "?" characters in a string or regex.
@@ -121,24 +130,17 @@ export class RedirectsMiddleware extends MiddlewareBase {
 
           // Redirect pattern matches the full incoming URL with query string present
           matchedQueryString = [
-            regexParser(redirect.pattern).test(`${normalizedPath}${targetQS}`),
-            regexParser(redirect.pattern).test(`/${locale}${normalizedPath}${targetQS}`),
+            regexParser(redirect.pattern).test(`${localePath}${incomingQS}`),
+            regexParser(redirect.pattern).test(`${normalizedPath}${incomingQS}`),
           ].some(Boolean)
-            ? targetQS
+            ? incomingQS
             : undefined;
-
           // Save the matched query string (if found) into the redirect object
           redirect.matchedQueryString = matchedQueryString || '';
-          debug.redirects('All info: %o', {
-            matchedQueryString,
-            patern: redirect.pattern,
-            targetURL,
-            targetQS,
-          });
           return (
             !!(
-              regexParser(redirect.pattern).test(targetURL) ||
-              regexParser(redirect.pattern).test(`/${req.nextUrl.locale}${targetURL}`) ||
+              regexParser(redirect.pattern).test(`/${req.nextUrl.locale}${incomingURL}`) ||
+              regexParser(redirect.pattern).test(incomingURL) ||
               matchedQueryString
             ) && (redirect.locale ? redirect.locale.toLowerCase() === locale.toLowerCase() : true)
           );
@@ -193,13 +195,14 @@ export class RedirectsMiddleware extends MiddlewareBase {
 
       // Find the redirect from result of RedirectService
       const existsRedirect = await this.getExistsRedirect(req, site.name);
-      debug.redirects('Existing redirect: %o', { existsRedirect });
 
       if (!existsRedirect) {
         debug.redirects('skipped (redirect does not exist)');
 
         return response;
       }
+
+      debug.redirects('Matched redirect rule: %o', { existsRedirect });
 
       // Find context site language and replace token
       if (